tests/wpt/web-platform-tests/IndexedDB/idbtransaction-objectStore-exception-order.html

<!doctype html>
<meta charset=utf-8>
<title>IndexedDB: IDBTransaction objectStore() Exception Ordering</title>
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbtransaction-objectstore">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<script>

indexeddb_test(
  (t, db) => {
    const store = db.createObjectStore('s');
  },
  (t, db) => {
    const tx = db.transaction('s');
    tx.oncomplete = t.step_func(() => {
        assert_throws('InvalidStateError', () => { tx.objectStore('nope'); },
                      '"finished" check (InvalidStateError) should precede ' +
                      '"name in scope" check (NotFoundError)');
      t.done();
    });
  },
  'IDBTransaction.objectStore exception order: InvalidStateError vs. NotFoundError'
);

</script>
